pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "i2c"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.45.0",
    "Pillow>=10.0.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "accelerate>=0.26.0",
    "bitsandbytes>=0.43.0",
    "qwen-vl-utils>=0.0.8",
]
